Move cache to mono source set
