--- app/shared/pages/cache-manage/common/package.kt
@@ -1,1 +1,0 @@
-package me.him188.ani.app.pages.cache.manage
