+++ app/shared/src/androidMain/kotlin/ui/cache/CacheItem.android.kt
@@ -1,4 +1,4 @@
-package me.him188.ani.app.pages.cache.manage
+package me.him188.ani.app.ui.cache
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
