+++ app/shared/src/androidMain/kotlin/ui/subject/collection/EpisodeProgress.android.kt
@@ -6,7 +6,7 @@ import androidx.compose.runtime.remember
 import androidx.compose.ui.tooling.preview.PreviewLightDark
 import me.him188.ani.app.data.media.EpisodeCacheStatus
 import me.him188.ani.app.data.subject.PackedDate
-import me.him188.ani.app.pages.cache.manage.testMediaCache1
+import me.him188.ani.app.ui.cache.testMediaCache1
 import me.him188.ani.app.ui.foundation.ProvideCompositionLocalsForPreview
 import me.him188.ani.app.ui.subject.collection.progress.EpisodeProgressDefaults
 import me.him188.ani.app.ui.subject.collection.progress.EpisodeProgressDialog
