+++ app/shared/src/commonMain/kotlin/ui/cache/package.kt
@@ -1,0 +1,1 @@
+package me.him188.ani.app.ui.cache
